Fixing quality, moving things aroung to correct places
