+++ frontend/src/registration/components/index.ts
@@ -2,4 +2,3 @@ export * from './RegistrationContainer';
 export * from './InstanceSetupPage';
 export * from './DomainInputPage';
 export * from './RegistrationPage';
-export * from './InstitutionalAccountHero';
