+++ frontend/src/ui/components/InstitutionalAccountHero/styles.scss
