+++ frontend/src/ui/components/index.ts
@@ -3,5 +3,5 @@ export * from './Footer';
 export * from './Header';
 export * from './Main';
 export * from './StepBar';
-
 export * from './DomainSuccessJumbotron';
+export * from './InstitutionalAccountHero';
